Improved process of 'not in' and fixed volume queries
This commit adds the possibility of use many "not in" concatenated one
after the other. Until now, it was not possible to do a query in the
form of "A not in B not in C not in D". This commit also fixes a problem
where the masks, which are supposed to be boolean matrices, where being
represented using integers.