Save all seed masks as inclusion masks
Now we force the seed masks to be used as inclusion masks. This is
because the seed masks are created using "endpoints_in". If we have
two "endpoints_in", then we are expecting streamlines to connect the two
endpoints. This can only be partially inforced by forcing the seed masks
to be also inclusion masks.